Add files via upload
Event binding Api that tracks the mouse and displays an array of images changing the one displayed based on user events. For the application to work properly you should load 2 images into the src/assets folder of you angular application and make the proper changes to event.component in regards to your images names and path.